+++ src/settings.js
@@ -1039,6 +1039,10 @@ const SETTINGS = {
             type: Boolean,
             default: false,
           },
+          "munching-policy-size-flat-scale": {
+            type: Boolean,
+            default: true,
+          },
           "munching-policy-use-source-filter": {
             type: Boolean,
             default: false,